# hexToString  optimization