simple optimize hexToString..
hexToRune is copy from encoding/json

benchmark
```
func Benchmark_hexToString(b *testing.B) {
	for i := 0; i < b.N; i++ {
		hexToString([]byte("0001F680"), 8)
	}
}
```

```
goos: darwin
goarch: arm64
pkg: tstgo
Benchmark_hexToString
Benchmark_hexToString-8    	30278191	        39.64 ns/op	      16 B/op	       2 allocs/op
Benchmark_hexToString2
Benchmark_hexToString2-8   	134665851	         8.828 ns/op	       0 B/op	       0 allocs/op
```