@kkHAIKE 2021-07-21
sorry.. hexToString2 test is new function, 8.828 ns/op, 4.5x faster than old one...