@pelletier 2021-07-21
Doh! I'm sorry, I somehow didn't see the benchmark in the PR description.

Pushed a commit to benchmark using `parser.parseBasicString` instead of `hexToString`, as it should be closer to what the user experiences. This made me realize that this function was only used in basic (multiline) string parsing, with its result always appended to a string buffer. As a result I changed the signature to return the rune directly, which removes some allocation. Then I inline `hexToRune` to return an error directly from the loop to scope the returned error to the specific invalid character, which should help the user see where the problem is.

---

Benchmarks:

### Baseline vs new hex2string implementation

```
name                           old time/op    new time/op     delta
ParseBasicStringWithUnicode/4     413ns ± 0%      259ns ± 0%  -37.37%  (p=0.000 n=9+9)
ParseBasicStringWithUnicode/8     353ns ± 0%      246ns ± 0%  -30.35%  (p=0.000 n=10+10)

name                           old speed      new speed       delta
ParseBasicStringWithUnicode/4  91.9MB/s ± 0%  146.8MB/s ± 0%  +59.65%  (p=0.000 n=9+9)
ParseBasicStringWithUnicode/8   119MB/s ± 0%    171MB/s ± 0%  +43.57%  (p=0.000 n=10+10)

name                           old alloc/op   new alloc/op    delta
ParseBasicStringWithUnicode/4      112B ± 0%        88B ± 0%  -21.43%  (p=0.000 n=10+10)
ParseBasicStringWithUnicode/8     96.0B ± 0%      80.0B ± 0%  -16.67%  (p=0.000 n=10+10)

name                           old allocs/op  new allocs/op   delta
ParseBasicStringWithUnicode/4      13.0 ± 0%        7.0 ± 0%  -46.15%  (p=0.000 n=10+10)
ParseBasicStringWithUnicode/8      9.00 ± 0%       5.00 ± 0%  -44.44%  (p=0.000 n=10+10)
```

### Baseline vs direct hex2rune call

```
name                           old time/op    new time/op     delta
ParseBasicStringWithUnicode/4     413ns ± 0%      146ns ± 0%   -64.62%  (p=0.000 n=9+10)
ParseBasicStringWithUnicode/8     353ns ± 0%      172ns ± 0%   -51.35%  (p=0.000 n=10+9)

name                           old speed      new speed       delta
ParseBasicStringWithUnicode/4  91.9MB/s ± 0%  259.8MB/s ± 0%  +182.67%  (p=0.000 n=9+10)
ParseBasicStringWithUnicode/8   119MB/s ± 0%    244MB/s ± 0%  +105.52%  (p=0.000 n=10+9)

name                           old alloc/op   new alloc/op    delta
ParseBasicStringWithUnicode/4      112B ± 0%        64B ± 0%   -42.86%  (p=0.000 n=10+10)
ParseBasicStringWithUnicode/8     96.0B ± 0%      64.0B ± 0%   -33.33%  (p=0.000 n=10+10)

name                           old allocs/op  new allocs/op   delta
ParseBasicStringWithUnicode/4      13.0 ± 0%        1.0 ± 0%   -92.31%  (p=0.000 n=10+10)
ParseBasicStringWithUnicode/8      9.00 ± 0%       1.00 ± 0%   -88.89%  (p=0.000 n=10+10)
```

### Side by side

```
name \ time/op                 baseline.txt   branch.txt      branch-2.txt
ParseBasicStringWithUnicode/4     413ns ± 0%      259ns ± 0%      146ns ± 0%
ParseBasicStringWithUnicode/8     353ns ± 0%      246ns ± 0%      172ns ± 0%

name \ speed                   baseline.txt   branch.txt      branch-2.txt
ParseBasicStringWithUnicode/4  91.9MB/s ± 0%  146.8MB/s ± 0%  259.8MB/s ± 0%
ParseBasicStringWithUnicode/8   119MB/s ± 0%    171MB/s ± 0%    244MB/s ± 0%

name \ alloc/op                baseline.txt   branch.txt      branch-2.txt
ParseBasicStringWithUnicode/4      112B ± 0%        88B ± 0%        64B ± 0%
ParseBasicStringWithUnicode/8     96.0B ± 0%      80.0B ± 0%      64.0B ± 0%

name \ allocs/op               baseline.txt   branch.txt      branch-2.txt
ParseBasicStringWithUnicode/4      13.0 ± 0%        7.0 ± 0%        1.0 ± 0%
ParseBasicStringWithUnicode/8      9.00 ± 0%       5.00 ± 0%       1.00 ± 0%
```