@pelletier 2021-07-21
I'm going to merge this as is given the significant improvement. Thank you so much for your patch and your help in improving go-toml!